requirements.txt
tensorflow
streamlit
numpy
opencv
Pillow